app/requirements.txt

Django
django-elasticsearch-dsl==7.3
django-elasticsearch-dsl-drf==0.22.5
django-environ==0.10.0
django-nine==0.2.7
django-rest-swagger==2.2.0
djangorestframework==3.14.0
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==5.2.2
elasticsearch-dsl
drf-yasg==1.21.6
drf-yasg2==1.19.4
elasticsearch==7.17.9
elasticsearch-dsl==7.4.1
django-dotenv
gunicorn==20.1.0

psycopg2-binary==2.9.6
dj-database-url
whitenoise
Brotli